% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyNamesID.R
\name{getSurveyNamesID}
\alias{getSurveyNamesID}
\title{Get survey names and IDs from Qualtrics.}
\usage{
getSurveyNamesID(username, token, format)
}
\arguments{
\item{username}{Your username from Qualtrics.
Defaults to 'username' from qualtricsAuth function 
(it has to be ran and loaded first).}

\item{token}{Your token from Qualtrics.
Defaults to 'token' from qualtricsAuth function 
(it has to be ran and loaded first).}

\item{format}{Valid options include \code{XML}, \code{JSON}, \code{HTML}, 
\code{JSON}; default is \code{CSV}.}
}
\value{
This function returns a data frame.
}
\description{
This function brings in your survey names and IDs from Qualtrics.
}
\details{
You can find your username and token in your account settings. 
Alternatively, you can use the 'qualtricsAuth' function to store a file 
with your username and token.
}
\examples{
\dontrun{

# Establish qualtricsAuth file first #

qualtricsAuth("username", "token") # You must provide these from Qualtrics.

load("file/location/qualtricsAuthInfo.RData")

getSurveyNamedsID()

# Without qualtricsAuth #

getSurveyNamedsID(username = "qualtricsUser@email.address#brand", 
                    token = "tokenString", format = "XML")
}
}
\author{
Seth Berry
}

