% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMaker.R
\name{groupMaker}
\alias{groupMaker}
\title{Assigns students to teams with 0/1 programming.}
\usage{
groupMaker(groupData, groupNumber, groupSizeMin, groupSizeMax)
}
\arguments{
\item{groupData}{Data from online student photos.}

\item{groupNumber}{How many groups you want.}

\item{groupSizeMin}{The minimum number of students per group.}

\item{groupSizeMax}{The maximum number of students per group.}
}
\value{
This will return a data frame with students and team assignments.
}
\description{
This function takes a vector of student names and creates groups.
}
\details{
The
}
\examples{
\dontrun{

studentNames <- c("Thomas", "Dirk", "John", "Brian", "Jenny", "Michael", 
"Bill", "Martin", "Douglas", "Hadley")

groupMaker(groupData = studentNames, groupNumber = 3, 
groupSizeMin = 3, groupSizeMax = 4)

}
}
\author{
Seth Berry
}
