% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQualtricsData.R
\name{importQualtricsData}
\alias{importQualtricsData}
\title{Imports data from Qualtrics into R.}
\usage{
importQualtricsData(username, token, surveyID, dropExtra, dropFirstRow, skip)
}
\arguments{
\item{username}{Your username from Qualtrics. Defaults to 'username' from
qualtricsAuth function (it has to be ran and loaded first).}

\item{token}{Your token from Qualtrics. Defaults to 'token' from
qualtricsAuth function (it has to be ran and loaded first).}

\item{surveyID}{The survey ID from Qualtrics.}

\item{dropExtra}{Logical (default: FALSE) indicating if you want the first 10
columns dropped. This is often useless data, but sometimes you might want
to keep the date.}

\item{dropFirstRow}{Logical (default: FALSE) indicating if you want the first
row dropped after read}

\item{skip}{(default: 0) passed into read.csv to determine the number of
rows to skip on read}
}
\value{
This function returns a data frame.
}
\description{
This function imports data from Qualtrics into R.
}
\details{
You can find your username, token, and survey ID in your account
settings.
Alternatively, you can use the 'qualtricsAuth' function to store a file with
your username and token, and then use the 'surveyNamesID' function to pull
all of the survey names into your R session.
}
\examples{
\dontrun{

# Establish qualtricsAuth file first #

qualtricsAuth("username", "token") # You must provide these from Qualtrics.

load("file/location/qualtricsAuthInfo.RData")

importQualtricsData(username, token, surveyID = "idString")

# Without qualtricsAuth #

importQualtricsData(username = "qualtricsUser@email.address#brand", 
                    token = "tokenString", surveyID = "idString", 
                    dropExtra = TRUE, skip = NULL, dropFirstRow = FALSE)
}
}
\author{
Seth Berry
}
