% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQualtricsDataV3.R
\name{importQualtricsDataV3}
\alias{importQualtricsDataV3}
\title{Imports data from Qualtrics into R using API V3.}
\usage{
importQualtricsDataV3(token, dataCenter, surveyID)
}
\arguments{
\item{token}{Your token from Qualtrics. Defaults to 'token' from
qualtricsAuth function (it has to be ran and loaded first).}

\item{dataCenter}{The Qualtrics data center for your account. If you don't 
know it, you can use "ca1". If that is not your data center, it will just
take a little longer to get your data.}

\item{surveyID}{The survey ID from Qualtrics.}
}
\value{
This function returns a data table from data.table.
}
\description{
This function imports data from Qualtrics into R using V3 API.
}
\details{
You can find your token and survey ID in your account
settings.
Alternatively, you can use the 'qualtricsAuth' function to store a file with
your username and token, and then use the 'surveyNamesID' function to pull
all of the survey names into your R session.
}
\examples{
\dontrun{

# Establish qualtricsAuth file first #

qualtricsAuth("username", "token") # You must provide these from Qualtrics.

load("file/location/qualtricsAuthInfo.RData")

dataTest <- importQualtricsDataV3(token, dataCenter, surveyID)

# Without qualtricsAuth #

dataTest <- importQualtricsDataV3(token = "yourAPIToken",
                            dataCenter = "ca1", surveyID = "yourSurveyID")
}
}
\author{
Seth Berry
}
