% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMatrix.R
\name{insertMatrix}
\alias{insertMatrix}
\title{Insert a matrix question}
\usage{
insertMatrix()
}
\description{
This function inserts a matrix question ([[Question:Matrix]])
for Advanced Format text files.
}
\details{
This function works best as an addin for RStudio (v0.99.878 or newer).
}
\examples{
\dontrun{
insertMatrix()
}
}
\author{
Seth Berry
}
