% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMultipleChoice.R
\name{insertMultipleChoice}
\alias{insertMultipleChoice}
\title{Insert a multiple choice question}
\usage{
insertMultipleChoice()
}
\description{
This function inserts a multiple choice question ([[Question:MC]])
for Advanced Format text files.
}
\details{
This function works best as an addin for RStudio (v0.99.878 or newer).
}
\examples{
\dontrun{
insertMultipleChoice()
}
}
\author{
Seth Berry
}

