% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPageBreak.R
\name{insertPageBreak}
\alias{insertPageBreak}
\title{Insert a page break}
\usage{
insertPageBreak()
}
\description{
This function inserts a page break question ([[PageBreak]])
for Advanced Format text files.
}
\details{
This function works best as an addin for RStudio (v0.99.878 or newer).
}
\examples{
insertPageBreak()
}
\author{
Seth Berry
}

