% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertTextEntry.R
\name{insertTextEntry}
\alias{insertTextEntry}
\title{Insert a text entry question}
\usage{
insertTextEntry()
}
\description{
This function inserts a text entry question ([[Question:TE]])
for Advanced Format text files.
}
\details{
This function works best as an addin for RStudio (v0.99.878 or newer).
}
\examples{
\dontrun{
insertTextEntry()
}
}
\author{
Seth Berry
}

