% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{q_read}
\alias{q_read}
\title{Read qualtrics file as a tibble}
\usage{
q_read(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded. Remote gz files can also be automatically downloaded &
   decompressed.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}
}
\value{
tibble. Column names derived using \code{\link{q_readv}}
}
\description{
Reads in a qualtrics file and handles formatting of header rows. Variable
names contained in the first two rows are read via \code{\link{q_readv}}. The
remainder of the data set is then read in, using the formatted variable names
as column names.
}

