% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{q_readv}
\alias{q_readv}
\title{Read variable information from a qualtrics file into a tibble of variable
names and labels}
\usage{
q_readv(file, drop_empty = TRUE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded. Remote gz files can also be automatically downloaded &
   decompressed.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}
}
\value{
tibble with variable (`var`) and corresponding label columns.
}
\description{
Reads in the first two lines of a Qualtrics Survey file. The first line
contains the variable names and the second contains the complete labels, such
as question text. The minor exception are the default Qualtrics variables,
which are named V1, V2, and so on. The variable name for these are stored in
the second row. Thus, for these, the label is substituted as the variable
name.
}

