% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{q_tidyloops}
\alias{q_tidyloops}
\title{Tidy columns generated in loop and merge blocks.}
\usage{
q_tidyloops(data, key_col = ".loop", value_col = ".dat",
  pattern = "\\\\([0-9]+\\\\)$")
}
\arguments{
\item{data}{A data frame}

\item{key_col}{Strings giving names of key and value columns to
create.}

\item{pattern}{character string containing a \link{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr} and \code{gregexpr}.}
}
\description{
Convenience function that takes a data frame and applies
\code{\link[tidyr]{gather}} to columns ending in `(number)`, which is how
loop and merge data is outputted. By default, it returns key-value results as
.loop and .dat
}

