\name{qualtricsAuth}
\alias{qualtricsAuth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
qualtricsAuth
}
\description{
This function stores your username and token in a file for easier access.
}
\usage{
qualtricsAuth(username, token)
}
\arguments{
  \item{username}{
Your username from Qualtrics.
}
  \item{token}{
Your token from Qualtrics
}
}

\author{
Seth Berry
}
\examples{

### Not Run ###

qualtricsAuth(qualtricsUser@email.address#brand, tokenString)

}

\keyword{Qualtrics}
