\name{qualtricsR-package}
\alias{qualtricsR-package}
\alias{qualtricsR}
\docType{package}
\title{
Functions for interacting with the Qualtrics REST API
}
\description{
This package has a number of functions for interacting with the Qualtrics REST API (exporting data, importing surveys, etc.). It also has some useful add-ins for creating 'Advanced Format' text files.
}
\details{
\tabular{ll}{
Package: \tab qualtricsR\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2015-01-06\cr
License: \tab MIT + file LICENSE\cr
}
}
\author{
Seth Berry

Maintainer: <seth.berry@nd.edu>
}
\references{
}
\keyword{ Qualtrics }
\seealso{
}
\examples{
}
