% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentSummary.R
\name{studentSummary}
\alias{studentSummary}
\title{Creates an Advanced Format text file from csv.}
\usage{
studentSummary(studentData)
}
\arguments{
\item{completeSurveyDataFrame}{The complete data frame containing the survey. 
Can be created as a data frame or read in from a file. Should contain 
variables called "question" (contains the question text), "responseOptions" (
contains the response option for each question, separated by some delimiter), 
and optionally "id" (the ID used in Qualtrics).}
}
\value{
This function returns a data frame of student scores.
}
\description{
This function will summarize student scores from Qualtrics data.
}
\details{
Provides summary stats for individual students.
}
\examples{
\dontrun{
questions = data.frame(question = c("I enjoy coding.",
                                    "To what extent do you hate or love R?", "Done?"),
                       responseOptions = c("No;Yes", 
                                           "Strongly hate;Hate;Neither;Love;Strongly love", 
                                           "No;Maybe;Yes"),
                       id = c("enjoyCode", "hateLoveR", "done"), 
                       stringsAsFactors = FALSE)


studentSummary(completeSurveyDataFrame = questions, roSeparator = ";", 
             pageBreakEvery = 2)
}
}
\author{
Seth Berry
}
