% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricsSurveyWriter.R
\name{surveyWriter}
\alias{surveyWriter}
\alias{createAdvancedFormat}
\title{Creates an Advanced Format text file from csv.}
\usage{
surveyWriter(completeSurveyDataFrame, roSeparator, pageBreakEvery,
                          outputFileName)
}
\arguments{
\item{completeSurveyDataFrame}{The complete data frame containing the survey. 
Can be created as a data frame or read in from a file. Should contain 
variables called "question" (contains the question text), "responseOptions" (
contains the response option for each question, separated by some delimiter), 
and optionally "questionID" (the ID used in Qualtrics).}

\item{roSeparator}{Character (default: ";") The character used for separating
the response options.}

\item{pageBreakEvery}{(default: 0) How many questions on a page before 
inserting a page break.}

\item{outputFileName}{(default: "surveyOut.txt") the name and location for
the outputted text file.}
}
\value{
This function returns a text file formatted for Qualtrics.
}
\description{
This function creates a text file for import into Qualtrics.
}
\details{
You can create the data frame or read in the survey from csv/xlsx.
}
\examples{
\dontrun{
questions = data.frame(question = c("I enjoy coding.",
                                    "To what extent do you hate or love R?", "Done?"),
                       responseOptions = c("No;Yes", 
                                           "Strongly hate;Hate;Neither;Love;Strongly love", 
                                           "No;Maybe;Yes"),
                       questionID = c("enjoyCode", "hateLoveR", "done"), 
                       stringsAsFactors = FALSE)


surveyWriter(completeSurveyDataFrame = questions, roSeparator = ";", 
             pageBreakEvery = 2)
}
}
\author{
Seth Berry
}
