% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamEmbeddedMaker.R
\name{teamEmbeddedMaker}
\alias{teamEmbeddedMaker}
\title{Creates crossings for each group to upload into Qualtrics.}
\usage{
teamEmbeddedMaker(teamAssignment)
}
\arguments{
\item{teamAssignmentData}{A data frame with student names (called "name") and
teams (called "teams").}
}
\value{
This will return a data frame with the crossing of students for each
group.
}
\description{
This function takes a vector of student names and creates groups.
}
\details{
This is most useful for creating embedded data fields for Qualtrics. 
It is best to use the object returned from the groupMaker function.
}
\examples{
\dontrun{

teamAssignment <- data.frame(name = c("Thomas", "Dirk", "John", "Brian", "Jenny", "Michael", 
"Bill", "Martin", "Douglas", "Hadley"), team = c(rep(1, 3), rep(2, 4), rep(3, 3)))

teamEmbeddedMaker(teamAssignment)

}
}
\author{
Seth Berry
}
