% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQualtricsSurvey.R
\name{importQualtricsSurvey}
\alias{importQualtricsSurvey}
\title{Import Advanced Format survey to Qualtrics}
\usage{
importQualtricsSurvey(username, token, surveyName, fileLocation)
}
\arguments{
\item{username}{Your username from Qualtrics.
Defaults to 'username' from qualtricsAuth function 
(it has to be ran and loaded first).}

\item{token}{Your token from Qualtrics.
Defaults to 'token' from qualtricsAuth function 
(it has to be ran and loaded first).}

\item{surveyName}{What you want your survey to be named within Qualtrics.}

\item{fileLocation}{Where your survey file is stored locally.}
}
\description{
This function imports an Advanced Format text file from R to Qualtrics.
}
\details{
You can find your username and token in your account settings. 
Alternatively, you can use the 'qualtricsAuth' function to store a file
 with your username and token.
}
\examples{
\dontrun{

# Establish qualtricsAuth file first #

qualtricsAuth("username", "token") # You must provide these from Qualtrics.

load("file/location/qualtricsAuthInfo.RData")

importQualtricsSurvey(username = username, token = token,
                      surveyName = "yourSurveyName",
                      fileLocation = "folder/location/yourSurveyName.txt")

# Without qualtricsAuth #

importQualtricsSurvey(username = "qualtricsUser@email.address#brand", 
                      token = "tokenString", surveyName = "yourSurveyName",
                      fileLocation = "folder/location/yourSurveyName.txt")
}
}
\author{
Seth Berry
}

