% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQualtricsSurvey.R
\name{importQualtricsSurvey}
\alias{importQualtricsSurvey}
\title{Import Advanced Format survey to Qualtrics}
\usage{
importQualtricsSurvey(username, token, format, surveyName, 
inputFormat, fileLocation)
}
\arguments{
\item{username}{Your username from Qualtrics.  Defaults to 'username' from qualtricsAuth 
function (it has to be ran and loaded first).}

\item{token}{Your token from Qualtrics.  Defaults to 'token' from qualtricsAuth 
function (it has to be ran and loaded first).}

\item{format}{Valid options include \code{XML}, \code{JSON}, \code{JSONP}; 
default is \code{XML}.}

\item{surveyName}{What you want your survey to be named within Qualtrics.}

\item{inputFormat}{The format of the file that you are importing. 
Possible values include: \code{TXT}, \code{QSF}, \code{DOC}, or \code{MSQ}; 
"TXT" is going to be a common file type if you are creating 
"Advanced Format" surveys. Defaults to "TXT".}

\item{fileLocation}{Where your survey file is stored locally.}
}
\description{
This function imports an Advanced Format text file from R to Qualtrics.
}
\details{
You can find your username and token in your account settings. 
Alternatively, you can use the 'qualtricsAuth' function to store a file
 with your username and token.
}
\examples{
\dontrun{

# Establish qualtricsAuth file first #

qualtricsAuth("username", "token") # You must provide these from Qualtrics.

load("file/location/qualtricsAuthInfo.RData")

importQualtricsSurvey(surveyName = "yourSurveyName", 
                      fileLocation = "folder/location/yourSurveyName.txt")

# Without qualtricsAuth #

importQualtricsSurvey(username = "qualtricsUser@email.address#brand", 
                      token = "tokenString", format = "XML", 
                      surveyName = "yourSurveyName", inputFormat = "TXT", 
                      fileLocation = "folder/location/yourSurveyName.txt")
}
}
\author{
Seth Berry
}

