% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricsAuth.R
\name{qualtricsAuth}
\alias{qualtricsAuth}
\title{Save Qualtrics Authentication Info.}
\usage{
qualtricsAuth(username, token, file)
}
\arguments{
\item{username}{Your username from Qualtrics.  Defaults to 'username' from qualtricsAuth 
function (it has to be ran and loaded first).}

\item{token}{Your token from Qualtrics.  Defaults to 'token' from qualtricsAuth 
function (it has to be ran and loaded first).}

\item{file}{Where you want the .RData file to be stored and what you want it to be names.
Defaults to "qualtricsAuthInfo.RData"}
}
\value{
This function creates an .RData file in you working directory.
}
\description{
This function stores your username and token in a file for easier access.
}
\details{
You can find your username and token in your account settings.
}
\examples{
\dontrun{

qualtricsAuth("qualtricsUser@email.address#brand", "tokenString", 
              file = "qualtricsAuthInfo.RData")
}
}
\author{
Seth Berry
}
